Refactor to new permissions system
