+++ packages/api/src/router/message/message.ts
@@ -5,16 +5,18 @@ import {
   router,
   publicProcedure,
 } from "~api/router/trpc";
-import {
-  protectedMessageFetch,
-  protectedMessageMutation,
-  protectedMessageMutationFetchFirst,
-} from "~api/utils/protected-procedures/message-editor-procedures";
+
 import { ignored_discord_account_router } from "../users/ignored-discord-accounts/ignored-discord-account";
 import { assertIsNotDeletedUser } from "~api/router/users/ignored-discord-accounts/ignored-discord-account";
 import { TRPCError } from "@trpc/server";
 import { findOrThrowNotFound } from "~api/utils/operations";
 import { discordAccountRouter, z_discord_account_public } from "../users/accounts/discord-accounts";
+import {
+  protectedFetch,
+  protectedMutation,
+  protectedMutationFetchFirst,
+} from "~api/utils/protected-procedures/base";
+import { assertCanEditMessage, assertCanEditMessages } from "~api/utils/permissions";
 
 const z_discord_image = z.object({
   url: z.string(),
@@ -47,10 +49,13 @@ const message_find_router = router({
     return findOrThrowNotFound(() => ctx.elastic.getMessage(input), "Message not found");
   }),
   byIdBulk: withDiscordAccountProcedure.input(z.array(z.string())).query(async ({ ctx, input }) => {
-    return protectedMessageFetch({
-      ctx,
-      author_id: input.map((i) => i),
+    return protectedFetch({
       fetch: () => ctx.elastic.bulkGetMessages(input),
+      permissions: (messages) =>
+        assertCanEditMessages(
+          ctx,
+          messages.map((m) => m.author_id)
+        ),
       not_found_message: "Message not found",
     });
   }),
@@ -72,17 +77,15 @@ const message_find_router = router({
 
 const message_crud_router = router({
   update: withDiscordAccountProcedure.input(z_message).mutation(async ({ ctx, input }) => {
-    return protectedMessageMutation({
-      ctx,
-      author_id: input.id,
+    return protectedMutation({
+      permissions: () => assertCanEditMessage(ctx, input.author_id),
       operation: () => ctx.elastic.updateMessage(input),
     });
   }),
   upsert: withDiscordAccountProcedure.input(z_message).mutation(async ({ ctx, input }) => {
     await assertIsNotDeletedUser(ctx, input.author_id);
-    return protectedMessageMutation({
-      ctx,
-      author_id: input.id,
+    return protectedMutation({
+      permissions: () => assertCanEditMessage(ctx, input.author_id),
       operation() {
         return ctx.elastic.upsertMessage(input);
       },
@@ -97,9 +100,12 @@ const message_crud_router = router({
         .byIdMany([...author_ids]);
       const ignored_account_ids = new Set(ignored_accounts.map((i) => i.id));
       const filtered_messages = input.filter((msg) => !ignored_account_ids.has(msg.author_id));
-      const status = await protectedMessageMutation({
-        ctx,
-        author_id: filtered_messages.map((msg) => msg.author_id),
+      const status = await protectedMutation({
+        permissions: () =>
+          assertCanEditMessages(
+            ctx,
+            filtered_messages.map((m) => m.author_id)
+          ),
         operation() {
           return ctx.elastic.bulkUpsertMessages(filtered_messages);
         },
@@ -113,20 +119,18 @@ const message_crud_router = router({
       return filtered_messages;
     }),
   delete: withDiscordAccountProcedure.input(z.string()).mutation(async ({ ctx, input }) => {
-    return protectedMessageMutationFetchFirst({
-      ctx,
-      getAuthorId: (data) => data.author_id,
+    return protectedMutationFetchFirst({
       fetch: () => message_find_router.createCaller(ctx).byId(input),
       operation: () => ctx.elastic.deleteMessage(input),
       not_found_message: "Message not found",
+      permissions: (message) => assertCanEditMessage(ctx, message.author_id),
     });
   }),
   deleteByThreadId: withDiscordAccountProcedure
     .input(z.string())
     .mutation(async ({ ctx, input }) => {
-      return protectedMessageMutationFetchFirst({
-        ctx,
-        getAuthorId: (data) => data.author_id,
+      return protectedMutationFetchFirst({
+        permissions: (message) => assertCanEditMessage(ctx, message.author_id),
         fetch: () => message_find_router.createCaller(ctx).byId(input),
         operation: () => ctx.elastic.deleteMessagesByThreadId(input),
         not_found_message: "Message not found",
@@ -135,9 +139,12 @@ const message_crud_router = router({
   deleteBulk: withDiscordAccountProcedure
     .input(z.array(z.string()))
     .mutation(async ({ ctx, input }) => {
-      return protectedMessageMutationFetchFirst({
-        ctx,
-        getAuthorId: (data) => data.map((d) => d.author_id),
+      return protectedMutationFetchFirst({
+        permissions: (messages) =>
+          assertCanEditMessages(
+            ctx,
+            messages.map((m) => m.author_id)
+          ),
         fetch: () => message_find_router.createCaller(ctx).byIdBulk(input),
         operation: () => ctx.elastic.bulkDeleteMessages(input),
         not_found_message: "Message not found",
