+++ packages/api/src/router/server/server.ts
@@ -2,10 +2,8 @@ import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 import { mergeRouters, withUserServersProcedure, router } from "~api/router/trpc";
 import { upsert } from "~api/utils/operations";
-import {
-  protectedServerManagerFetch,
-  protectedServerManagerMutation,
-} from "~api/utils/protected-procedures/server-manager-procedures";
+import { assertCanEditServer } from "~api/utils/permissions";
+import { protectedFetch, protectedMutation } from "~api/utils/protected-procedures/base";
 
 export const z_server = z.object({
   id: z.string(),
@@ -36,33 +34,28 @@ export const z_server_upsert = z_server_create;
 
 const serverCreateUpdateRouter = router({
   create: withUserServersProcedure.input(z_server_create).mutation(({ ctx, input }) => {
-    return protectedServerManagerMutation({
-      ctx,
-      server_id: input.id,
+    return protectedMutation({
       operation: () => ctx.prisma.server.create({ data: input }),
+      permissions: () => assertCanEditServer(ctx, input.id),
     });
   }),
   update: withUserServersProcedure.input(z_server_update).mutation(({ ctx, input }) => {
-    return protectedServerManagerMutation({
-      ctx,
-      server_id: input.id,
+    return protectedMutation({
       operation: () => ctx.prisma.server.update({ where: { id: input.id }, data: input }),
+      permissions: () => assertCanEditServer(ctx, input.id),
     });
   }),
 });
 
 const serverFetchRouter = router({
   byId: withUserServersProcedure.input(z.string()).query(async ({ ctx, input }) => {
-    return protectedServerManagerFetch({
+    return protectedFetch({
       async fetch() {
         const server = await ctx.prisma.server.findUnique({ where: { id: input } });
         if (!server) throw new TRPCError({ code: "NOT_FOUND", message: "Server not found" });
         return server;
       },
-      getServerId(data) {
-        return data.id;
-      },
-      ctx,
+      permissions: () => assertCanEditServer(ctx, input),
       not_found_message: "Server not found",
     });
   }),
