+++ packages/api/src/router/users/ignored-discord-accounts/ignored-discord-account.ts
@@ -1,9 +1,10 @@
 import { z } from "zod";
-import { protectedUserOnlyMutation } from "~api/utils/protected-procedures/user-only";
 import { withDiscordAccountProcedure, router, publicProcedure } from "~api/router/trpc";
 import { findAllowNull, findOrThrowNotFound } from "~api/utils/operations";
 import type { Context } from "~api/router/context";
 import { TRPCError } from "@trpc/server";
+import { protectedMutation } from "~api/utils/protected-procedures/base";
+import { assertIsUser } from "~api/utils/permissions";
 
 export async function assertIsNotDeletedUser(ctx: Context, target_user_id: string) {
   const deleted_account = await findAllowNull(() =>
@@ -20,9 +21,8 @@ export async function assertIsNotDeletedUser(ctx: Context, target_user_id: strin
 
 export const ignored_discord_account_router = router({
   upsert: withDiscordAccountProcedure.input(z.string()).mutation(({ ctx, input }) => {
-    return protectedUserOnlyMutation({
-      ctx,
-      user_id: input,
+    return protectedMutation({
+      permissions: () => assertIsUser(ctx, input),
       operation: () =>
         ctx.prisma.ignoredDiscordAccount.upsert({
           where: { id: input },
@@ -52,9 +52,8 @@ export const ignored_discord_account_router = router({
     });
   }),
   stopIgnoring: withDiscordAccountProcedure.input(z.string()).mutation(({ ctx, input }) => {
-    return protectedUserOnlyMutation({
-      ctx,
-      user_id: input,
+    return protectedMutation({
+      permissions: () => assertIsUser(ctx, input),
       operation: () => ctx.prisma.ignoredDiscordAccount.delete({ where: { id: input } }),
     });
   }),
