--- packages/api/src/utils/protected-procedures/message-editor-procedures.ts
@@ -1,46 +1,0 @@
-import type { Context } from "~api/router/context";
-import { assertCanEditMessages } from "../permissions";
-import {
-  protectedFetch,
-  ProtectedFetchInputExtendable,
-  protectedMutation,
-  protectedMutationFetchFirst,
-  ProtectedMutationFetchFirstInputExtendable,
-  ProtectedMutationInputExtendable,
-} from "./base";
-
-export async function protectedMessageMutation<T>(
-  input: ProtectedMutationInputExtendable<T> & {
-    author_id: string | string[];
-    ctx: Context;
-  }
-) {
-  return protectedMutation({
-    ...input,
-    assertPermissions: () => assertCanEditMessages(input.ctx, input.author_id),
-  });
-}
-
-export async function protectedMessageFetch<T>(
-  input: ProtectedFetchInputExtendable<T> & {
-    author_id: string | string[];
-    ctx: Context;
-  }
-) {
-  return protectedFetch({
-    ...input,
-    assertPermissions: () => assertCanEditMessages(input.ctx, input.author_id),
-  });
-}
-
-export async function protectedMessageMutationFetchFirst<T, F>(
-  input: ProtectedMutationFetchFirstInputExtendable<T, F> & {
-    getAuthorId: (data: T) => string | string[];
-    ctx: Context;
-  }
-) {
-  return protectedMutationFetchFirst({
-    ...input,
-    assertPermissions: (data) => assertCanEditMessages(input.ctx, input.getAuthorId(data)),
-  });
-}
